requirements.txt
pandas==2.2.2
numpy==1.26.4
streamlit==1.37.1
configparser==7.0.0
google-generativeai==0.7.2
langchain==0.2.12
langchain-community==0.2.11
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
langchain-google-genai==1.0.8
chromadb==0.5.5
mne==1.7.1
aioredis==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.6.2
cffi==1.16.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.8
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==3.2.10
django-channels==0.7.0
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
gunicorn==22.0.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
intel-openmp==2021.4.0
mkl==2021.4.0
msgpack==1.0.8
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.3
packaging==24.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
tablib==3.6.1
tbb==2021.12.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.4.post2
